refactor(convex): dedupe cross-domain storage deletion in mediaStorage
