+++ convex/mediaStorage.ts
@@ -1,0 +1,47 @@
+import type { Id } from "./_generated/dataModel";
+import type { MutationCtx, QueryCtx } from "./_generated/server";
+
+export type StorageMetadataRecord = {
+  _id: Id<"_storage">;
+  _creationTime: number;
+  contentType?: string;
+  sha256: string;
+  size: number;
+};
+
+export async function getStorageMetadata(
+  ctx: QueryCtx | MutationCtx,
+  storageId: Id<"_storage">,
+): Promise<StorageMetadataRecord | null> {
+  return (await ctx.db.system.get(
+    "_storage",
+    storageId,
+  )) as StorageMetadataRecord | null;
+}
+
+export async function deleteStorageIfUnreferenced(
+  ctx: MutationCtx,
+  storageId: Id<"_storage">,
+): Promise<boolean> {
+  const [galleryRefs, audioRefs] = await Promise.all([
+    ctx.db
+      .query("galleryPhotos")
+      .withIndex("by_storageId", (q) => q.eq("storageId", storageId))
+      .take(1),
+    ctx.db
+      .query("audioTracks")
+      .withIndex("by_storageId", (q) => q.eq("storageId", storageId))
+      .take(1),
+  ]);
+  if (galleryRefs.length > 0 || audioRefs.length > 0) {
+    return false;
+  }
+
+  const storage = await getStorageMetadata(ctx, storageId);
+  if (!storage) {
+    return false;
+  }
+
+  await ctx.storage.delete(storageId);
+  return true;
+}
